fix(faqscreen corrigida apos reorganizaçao): correçao de paths de arquivos
